🐛 Fix issue with reading directory taking long
Removed dependency directory-tree as the library was designed to only get all subdirectories. Thus using this library not only lengthed reading the source directory, but also is not a functionality that is needed in scope of this project.

This library was replaced with our own implementation getDirectoryTree.ts where we can control the dive into subdirectories (which we do not want to do anyway). getDataXmlAsJson.ts and getDataXml.ts have been updated using getDirectoryTree.ts and corresponding types - minimal changes where made beyond changes expected types.