🐛 Fix reading XML in UTF-16 or UCS-2 encoding
data XML files from CHILI 4 were written in UCS-2 LE encoding.
getStringFromFile() will take that into account using chardet. Currently
on support for UTF-16 and UTF-8 are allowed. For other support,
something like Incov-lite would be needed to convert as node has limited
support for encodings.

getDataXmlAsJosn() was also updated to use no-try, as
getStringFromFile() can return an error if the file is not found, a
permission issue is recieved, or the file it not of type file.